Restructured navbar directory
