+++ src/components/Button.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import "./Button.css";
 
 const STYLES = ["btn--primary", "btn--submit", "btn--interact"];
