+++ src/components/navbar/Navbar.jsx
@@ -2,20 +2,9 @@ import { React, useEffect, useState } from "react";
 import { useMsal } from "@azure/msal-react";
 import "./Navbar.css";
 import "../../App.css";
-import UserView from "../navComponents/UserView";
-import { NavButton } from "../navComponents/NavButton";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faHouse,
-  faGear,
-  faPeopleGroup,
-  faLaptop,
-  faFile,
-  faPrint,
-  faHeadset,
-  faNewspaper,
-} from "@fortawesome/free-solid-svg-icons";
+import UserView from "./components/UserView";
 import { Button } from "../Button";
+import NavMenu from "./components/NavMenu";
 
 function Navbar() {
   const [selectedBtn, setSelectedBtn] = useState(null);
@@ -41,78 +30,7 @@ function Navbar() {
           <div className="userInfo">
             <UserView />
           </div>
-          <div id="nav-list" className="nav-btns">
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Home")}
-            >
-              <NavButton
-                name="Home"
-                icon={<FontAwesomeIcon className="fa-icon" icon={faHouse} />}
-                selected={selectedBtn}
-                linkto="/dashboard/home"
-              />
-            </div>
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Employees")}
-            >
-              <NavButton
-                name="Employees"
-                icon={
-                  <FontAwesomeIcon className="fa-icon" icon={faPeopleGroup} />
-                }
-                linkto="/dashboard/users"
-                selected={selectedBtn}
-              />
-            </div>
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Job Listings")}
-            >
-              <NavButton
-                name="Job Listings"
-                icon={
-                  <FontAwesomeIcon className="fa-icon" icon={faNewspaper} />
-                }
-                linkto="/dashboard/jobs"
-                selected={selectedBtn}
-              />
-            </div>
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Add a Device")}
-            >
-              <NavButton
-                name="Add a Device"
-                icon={<FontAwesomeIcon className="fa-icon" icon={faPrint} />}
-                linkto="/dashboard/adddevice"
-                selected={selectedBtn}
-              />
-            </div>
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Support")}
-            >
-              <NavButton
-                name="Support"
-                icon={<FontAwesomeIcon className="fa-icon" icon={faHeadset} />}
-                linkto="/dashboard/support"
-                selected={selectedBtn}
-              />
-            </div>
-            <div
-              className="btn-container"
-              onClick={() => setSelectedBtn("Settings")}
-            >
-              <NavButton
-                name="Settings"
-                icon={<FontAwesomeIcon className="fa-icon" icon={faGear} />}
-                linkto="/dashboard/settings"
-                selected={selectedBtn}
-              />
-            </div>
-          </div>
+          <NavMenu selection={setSelectedBtn} selected={selectedBtn} />
           <div className="logout-button">
             <Button
               id="logout"
