+++ src/components/navbar/components/UserView.scss
